move fakemysqldaemon into its own package
Since fakemysqldaemon pulls in fakemysqldb, which in turn pulls in
the "testing" package, it ended up adding the -test.xyz flags to
the vttablet binary.

Pulling fakemysqldaemon into its own package avoids this issue and is
better code hygiene since it removes test code from the build.
