+++ go/vt/vttablet/tabletmanager/binlog_players_test.go
@@ -30,7 +30,7 @@ import (
 	"github.com/youtube/vitess/go/vt/binlog/binlogplayer"
 	"github.com/youtube/vitess/go/vt/grpcclient"
 	"github.com/youtube/vitess/go/vt/key"
-	"github.com/youtube/vitess/go/vt/mysqlctl"
+	"github.com/youtube/vitess/go/vt/mysqlctl/fakemysqldaemon"
 	"github.com/youtube/vitess/go/vt/mysqlctl/tmutils"
 	"github.com/youtube/vitess/go/vt/topo"
 	"github.com/youtube/vitess/go/vt/topo/memorytopo"
@@ -262,7 +262,7 @@ func TestBinlogPlayerMapHorizontalSplit(t *testing.T) {
 	// create the BinlogPlayerMap on the local tablet
 	// (note that local tablet is never in the topology, we don't
 	// need it there at all)
-	mysqlDaemon := &mysqlctl.FakeMysqlDaemon{MysqlPort: 3306}
+	mysqlDaemon := &fakemysqldaemon.FakeMysqlDaemon{MysqlPort: 3306}
 	vtClientSyncChannel := make(chan *binlogplayer.VtClientMock)
 	bpm := NewBinlogPlayerMap(ts, mysqlDaemon, func() binlogplayer.VtClient {
 		return <-vtClientSyncChannel
@@ -448,7 +448,7 @@ func TestBinlogPlayerMapHorizontalSplitStopStartUntil(t *testing.T) {
 	// create the BinlogPlayerMap on the local tablet
 	// (note that local tablet is never in the topology, we don't
 	// need it there at all)
-	mysqlDaemon := &mysqlctl.FakeMysqlDaemon{MysqlPort: 3306}
+	mysqlDaemon := &fakemysqldaemon.FakeMysqlDaemon{MysqlPort: 3306}
 	vtClientSyncChannel := make(chan *binlogplayer.VtClientMock)
 	bpm := NewBinlogPlayerMap(ts, mysqlDaemon, func() binlogplayer.VtClient {
 		return <-vtClientSyncChannel
@@ -642,7 +642,7 @@ func TestBinlogPlayerMapVerticalSplit(t *testing.T) {
 	// (note that local tablet is never in the topology, we don't
 	// need it there at all)
 	// The schema will be used to resolve the table wildcards.
-	mysqlDaemon := &mysqlctl.FakeMysqlDaemon{
+	mysqlDaemon := &fakemysqldaemon.FakeMysqlDaemon{
 		MysqlPort: 3306,
 		Schema: &tabletmanagerdatapb.SchemaDefinition{
 			DatabaseSchema: "",
